Program now read questions in from yaml file
Firstly, I separated the program into multiple files similar to how I
did with my go project.

Then, I changed the program such that it now reads in from a yaml file.
It then creates a quiz object which in turn creates subject and question
objects, which are then used to ask the questions and check for answers.
I have also changed how answers are checked. The program now iterates
over each answer and casts it to a string, before it is checked,
preventing type errors e.g. '1919' != 1919.