Started work on port to rust
I have started work on porting the project to rust, since I have heard a
lot about rust and wanted to see what I thought of it. I have made
little progress so far and it is giving me flashbacks to my nightmares
with see, but I'll 'c' how I get on with it.

I may also port the project over to javascript as this may be easier to
run in a browser. Since one of the pupils in my year has made history
website where it could be hosted, this may make it easier to distribute.
However, their website has been made using WIX, so this may make adding
custom things, such as an interactive quiz, with user made questions,
harder.