Added schema validation
I used JSON schema to allow the program to validate the yaml file given
by the user, since yaml is just converted to json when read in. It will
check to make sure that it is compatible with the program, and if it is
not it will give the user an error message and use the default question
set.

At the moment, the only extra thing I can think that I could add would
be subtopics e.g. Treaty of Versailles for History, and Evolution for
Science. However, this could pose an issue as I would have to choose
whether the user had to make subtopics (probably), or if they could
chose to make them. The former may frustrate users that don't want to, or
don't know how to, split their question set into subtopics, and the latter
may be difficult to set up.